requirements.txt
pytorch-lightning==1.7.7
torchmetrics==0.11.2
lpips==0.1.4
matplotlib==3.5.3
omegaconf==2.2.3
opencv-python==4.7.0.68
trimesh==3.23.5
imageio==2.24.0
pysdf==0.1.8
chumpy
numpy==1.21
pymcubes==0.1.2